requirements.txt
django==3.2
djangorestframework==3.12.2
drf-yasg==1.20.0
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.7.0
model-bakery==1.3.1
drf-extra-fields==3.1.1